shuffled things around to get workspaceGroup registration in routes.json working
