+++ packages/framework/esm-extensions/src/index.ts
@@ -3,6 +3,7 @@ export * from './helpers';
 export * from './left-nav';
 export * from './modals';
 export * from './workspaces';
+export * from './workspaces2';
 export * from './render';
 export * from './store';
 export * from './types';
