+++ packages/framework/esm-styleguide/src/workspaces2/index.tsx
@@ -1,3 +1,2 @@
 export * from './workspace2.component';
 export * from './workspace2';
-export * from './store';
